backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
cloudinary==1.44.1
dj-database-url==3.1.2
Django==5.0.14
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
idna==3.11
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-decouple==3.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
reportlab==4.2.5
stripe==14.4.1
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.12.0
pyotp==2.9.0
qrcode==8.2
